+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
 - Added the possibility to specify engine parameters through the registry
   [ale-rt]
 
+- Allow to create and configure a database file by installing the `database` setup profile
+
 
 1.0.1 (2016-12-28)
 ------------------