+++ CHANGES.rst
@@ -8,13 +8,16 @@ Changelog
 - Added the possibility to specify engine parameters through the registry
   [ale-rt]
 
+- remove unneeded columns in ORM model (site_name, title, info) [fRiSi]
+
 - Implemented API for scheduling items as must-read for certain users.
   (see collective.contentrules.mustread for usage)
 
   This required new database columns. The provided upgrade step works for sqlite databases
-  but might need changes for mysql or postgres.
+  but might need changes for mysql or postgres. [fRiSi]
 
 - Allow to create and configure a database file by installing the `database` setup profile
+  [fRiSI]
 
 
 1.0.1 (2016-12-28)